requirements.txt
python
mysql-connector
Dotenv
